refactor : 패키지 관리 변경
