+++ src/main/java/blackjack/domain/participants/BettingResult.java
@@ -1,5 +1,6 @@
 package blackjack.domain.participants;
 
+import blackjack.domain.Profit;
 import java.util.LinkedHashMap;
 import java.util.Map;
