+++ src/main/java/blackjack/view/InputView.java
@@ -1,6 +1,5 @@
 package blackjack.view;
 
-import blackjack.domain.participants.Profit;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
