+++ src/test/java/blackjack/domain/ParticipantsTest.java
@@ -11,8 +11,6 @@ import blackjack.domain.participants.Name;
 import blackjack.domain.participants.Participants;
 import blackjack.domain.participants.Player;
 import blackjack.domain.participants.Players;
-import blackjack.domain.participants.BlackJackGameResult;
-import blackjack.domain.participants.State;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
