+++ src/test/java/blackjack/domain/PlayersTest.java
@@ -9,7 +9,6 @@ import blackjack.domain.participants.Player;
 import blackjack.domain.participants.Players;
 import java.util.List;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
